Add files via upload
The starting point is the lower-left corner, which is highlighted in blue. The point of the destination is highlighted in black. We can move one unit in each stage and we can only go to the destination through the green dots. The red dots are actually obstacles in the way that we have to somehow get around. The program is supposed to suggest the shortest route between origin and destination.